requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
attrs==24.3.0
bcrypt==4.2.0
blinker==1.8.2
botocore==1.35.49
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
filelock==3.16.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.3
Flask-WTF==1.2.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
phonenumbers==8.13.48
platformdirs==4.3.2
postmarker==1.0
propcache==0.2.1
PyJWT==2.9.0
pymongo==4.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
stripe==11.2.0
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.4
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.18.3
